Major Changes Made, Branching Out
I was able to string together a few functions to create a basic
implementation of what our shell would look like.
Use `gcc ...options *.c` to compile.

The Content Of Each Source File
===============================

- `append.c` contains functions that automatically reallocate memory
when a buffer is full. appendChar works on strings while appendStr works
on arrays of strings.

- `quote.c` contains functions use to find and group a quoted substring
while tokenizing.

- `line.c` contains functions used to get a line and tokenize it.
tokenizeLine returns a null-terminated array of strings.

- `find.c` contains functions for getting the value of an environment
variable, searching for a file in a directory, and most importantly,
searching for a command in $PATH

- `run.c` currently contains a single function (execute) which simulates
execution of a known program, but actually does nothing.

- `shell.h` contains the prototypes of all the functions created so far.

- `main.c` contains the main function which creates the shell.

Betty Checks ✅
Compiles With No Errors ✅
No Memory Leaks ✅