Created Separate Buffers For $$ and $?
Previously, they both used the same buffer and when
$$ and $? were both present on a line, the buffer
would hold the string representation of the last one
encountered.

With a pid of 55467 and an errno of 21, the following would happen

($) echo $$ $? $$ $? $$
55467 55467 55467 55467 55467
($) echo $$ $? $$ $? $?
21 21 21 21 21

I've created separate buffers for them in the state. Now, they do
not affect each other. Furthermore, the buffer that holds pid would
only be modified once (when the program starts) since the pid never
changes.